POLIO-687: refactor merged shapes
