+++ plugins/polio/js/src/pages/Calendar.js
@@ -238,7 +238,6 @@ const Calendar = ({ params }) => {
                                 campaigns={filteredCampaigns}
                                 loadingCampaigns={isLoading}
                                 isPdf={isPdf}
-                                currentDate={currentDate}
                             />
                         </Grid>
                     </Grid>
